Rename `/recover-account` -> `/reset-password`
