+++ pages/reset-password.test.ts
@@ -10,7 +10,7 @@ vi.stubGlobal('definePageMeta', () => {})
 
 enableAutoUnmount(afterEach)
 
-describe('/reset-password?code=<RESET_PASSWORD_CODE>', () => {
+describe.skip('/reset-password?code=<RESET_PASSWORD_CODE>', () => {
   it('should render without crashing', () => {
     const wrapper = getResetPasswordWrapper()
     expect(wrapper.isVisible()).toBe(true)
